# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusMart API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  email: string;
  phone: string;
  username: string;
  password: string;
  campus?: string;
}

export interface LoginRequest {
  emailOrPhone: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  admin: "admin",
} as const;

export interface User {
  id: string;
  email: string;
  phone?: string;
  username: string;
  campus?: string;
  avatarUrl?: string;
  role?: UserRole;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserProfile {
  id: string;
  email: string;
  phone?: string;
  username: string;
  campus?: string;
  avatarUrl?: string;
  rating?: number;
  totalSales?: number;
  totalListings?: number;
  createdAt?: string;
}

export interface UpdateProfileRequest {
  username?: string;
  phone?: string;
  campus?: string;
  avatarUrl?: string;
}

export type ProductCondition =
  (typeof ProductCondition)[keyof typeof ProductCondition];

export const ProductCondition = {
  new: "new",
  like_new: "like_new",
  good: "good",
  fair: "fair",
} as const;

export type ProductStatus = (typeof ProductStatus)[keyof typeof ProductStatus];

export const ProductStatus = {
  active: "active",
  sold: "sold",
  paused: "paused",
} as const;

export type ProductBadge = (typeof ProductBadge)[keyof typeof ProductBadge];

export const ProductBadge = {
  HOT: "HOT",
  NEW: "NEW",
  SALE: "SALE",
} as const;

export interface Product {
  id: string;
  title: string;
  description?: string;
  price: number;
  originalPrice?: number;
  category: string;
  condition?: ProductCondition;
  campus: string;
  images?: string[];
  stock?: number;
  status?: ProductStatus;
  badge?: ProductBadge;
  sellerId?: string;
  sellerUsername?: string;
  sellerAvatar?: string;
  sellerRating?: number;
  isWishlisted?: boolean;
  createdAt?: string;
}

export interface CreateProductRequest {
  title: string;
  description?: string;
  price: number;
  originalPrice?: number;
  category: string;
  condition: string;
  campus: string;
  images?: string[];
  stock?: number;
}

export interface ProductsResponse {
  products: Product[];
  total: number;
  page?: number;
  totalPages?: number;
}

export interface Category {
  id: string;
  name: string;
  icon?: string;
  count?: number;
}

export interface CartItem {
  id: string;
  productId: string;
  title: string;
  price: number;
  image?: string;
  quantity: number;
  sellerId?: string;
  sellerUsername?: string;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartRequest {
  productId: string;
  quantity: number;
}

export interface UpdateCartItemRequest {
  quantity: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  awaiting_payment: "awaiting_payment",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderItem {
  productId: string;
  title: string;
  price: number;
  quantity: number;
  image?: string;
  sellerUsername?: string;
}

export interface Order {
  id: string;
  orderId: string;
  items: OrderItem[];
  totalAmount: number;
  status: OrderStatus;
  deliveryAddress?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface CreateOrderRequest {
  deliveryAddress: string;
  paymentMethod?: string;
}

export type RoomType = (typeof RoomType)[keyof typeof RoomType];

export const RoomType = {
  bedsitter: "bedsitter",
  single: "single",
  one_bedroom: "one_bedroom",
  two_bedroom: "two_bedroom",
  hostel: "hostel",
} as const;

export interface Room {
  id: string;
  title: string;
  description?: string;
  type: RoomType;
  monthlyRent: number;
  campus: string;
  distanceToCampus?: string;
  images?: string[];
  amenities?: string[];
  available: boolean;
  landlordName?: string;
  landlordPhone?: string;
  createdAt?: string;
}

export interface CreateRoomRequest {
  title: string;
  description?: string;
  type: string;
  monthlyRent: number;
  campus: string;
  distanceToCampus?: string;
  images?: string[];
  amenities?: string[];
  landlordPhone?: string;
}

export interface FoodVendor {
  id: string;
  name: string;
  campus: string;
  bannerImage?: string;
  rating?: number;
  deliveryTime?: string;
  minOrder?: number;
  categories?: string[];
  isOpen?: boolean;
}

export interface FoodItem {
  id: string;
  vendorId: string;
  vendorName?: string;
  name: string;
  description?: string;
  price: number;
  image?: string;
  category: string;
  available?: boolean;
  campus?: string;
}

export interface WishlistToggleRequest {
  productId: string;
}

export interface WishlistToggleResponse {
  wishlisted: boolean;
  productId: string;
}

export type ListProductsParams = {
  category?: string;
  campus?: string;
  search?: string;
  sort?: ListProductsSort;
  page?: number;
  limit?: number;
  featured?: boolean;
};

export type ListProductsSort =
  (typeof ListProductsSort)[keyof typeof ListProductsSort];

export const ListProductsSort = {
  newest: "newest",
  price_asc: "price_asc",
  price_desc: "price_desc",
  popular: "popular",
} as const;

export type ListRoomsParams = {
  type?: string;
  campus?: string;
  minPrice?: number;
  maxPrice?: number;
};

export type ListFoodVendorsParams = {
  campus?: string;
};

export type ListFoodItemsParams = {
  vendorId?: string;
  category?: string;
  campus?: string;
};