# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CampusMart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user's cart
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartRequest"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Cart cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /cart/{itemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List user orders
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /food/vendors:
    get:
      operationId: listFoodVendors
      tags: [food]
      summary: List food vendors
      parameters:
        - name: campus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Vendors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodVendor"
  /food/items:
    get:
      operationId: listFoodItems
      tags: [food]
      summary: List food items
      parameters:
        - name: vendorId
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: campus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Food items list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodItem"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        phone:
          type: string
        username:
          type: string
        password:
          type: string
        campus:
          type: string
      required:
        - email
        - phone
        - username
        - password
    LoginRequest:
      type: object
      properties:
        emailOrPhone:
          type: string
        password:
          type: string
      required:
        - emailOrPhone
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        phone:
          type: string
        username:
          type: string
        campus:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
          enum: [student, admin]
        createdAt:
          type: string
      required:
        - id
        - email
        - username
...
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: number
      required:
        - items
        - total
        - itemCount
    CartItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        title:
          type: string
        price:
          type: number
        image:
          type: string
        quantity:
          type: number
        sellerId:
          type: string
        sellerUsername:
          type: string
      required:
        - id
        - productId
        - title
        - price
        - quantity
    AddToCartRequest:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: number
      required:
        - productId
        - quantity
    UpdateCartItemRequest:
      type: object
      properties:
        quantity:
          type: number
      required:
        - quantity
    Order:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, awaiting_payment, confirmed, shipped, delivered, cancelled]
        deliveryAddress:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderId
        - items
        - totalAmount
        - status
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        title:
          type: string
        price:
          type: number
        quantity:
          type: number
        image:
          type: string
        sellerUsername:
          type: string
      required:
        - productId
        - title
        - price
        - quantity
    CreateOrderRequest:
      type: object
      properties:
        deliveryAddress:
          type: string
        paymentMethod:
          type: string
      required:
        - deliveryAddress
...
    FoodVendor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        campus:
          type: string
        bannerImage:
          type: string
        rating:
          type: number
        deliveryTime:
          type: string
        minOrder:
          type: number
        categories:
          type: array
          items:
            type: string
        isOpen:
          type: boolean
      required:
        - id
        - name
        - campus
    FoodItem:
      type: object
      properties:
        id:
          type: string
        vendorId:
          type: string
        vendorName:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        image:
          type: string
        category:
          type: string
        available:
          type: boolean
        campus:
          type: string
      required:
        - id
        - vendorId
        - name
        - price
        - category
...