# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusMart API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  id: string;
  productId: string;
  title: string;
  price: number;
  image?: string;
  quantity: number;
  sellerId?: string;
  sellerUsername?: string;
}