# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusMart API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: string;
  email: string;
  phone?: string;
  username: string;
  campus?: string;
  avatarUrl?: string;
  role?: UserRole;
  createdAt?: string;
}